+++ src/components/Link/Link.tsx
@@ -2,15 +2,12 @@
  * @author Pavel Kalachev <pkalachev@usetech.com>
  */
 
-import React, { FC } from 'react';
+import React, { AnchorHTMLAttributes, FC } from 'react';
 import './Link.scss';
 
-interface LinkProps {
+interface LinkProps extends AnchorHTMLAttributes<HTMLAnchorElement> {
     title: string;
-    href?: string;
     role?: 'primary' | 'secondary' | 'danger';
-    rel?: string;
-    target?: '_blank' | '_parent' | '_self' | '_top';
     onClick?(): void;
 }