+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Changes:
 
 - Loader component add
 - Upload component callback fix, image size fix
+- Link component extends native link props
 
 ## 0.0.16 May 11, 2022